:truck: Moved the MocksApiCommunication trait to the traits folder in tests.
