+++ tests/Feature/Proxy/FileProxyTest.php
@@ -7,7 +7,7 @@ use MyParcelCom\ApiSdk\Authentication\AuthenticatorInterface;
 use MyParcelCom\ApiSdk\MyParcelComApi;
 use MyParcelCom\ApiSdk\MyParcelComApiInterface;
 use MyParcelCom\ApiSdk\Resources\Proxy\FileProxy;
-use MyParcelCom\ApiSdk\Tests\Feature\MocksApiCommunication;
+use MyParcelCom\ApiSdk\Tests\Traits\MocksApiCommunication;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\StreamInterface;
 use Symfony\Component\Cache\Simple\NullCache;
